veml7700

+++ drivers/iio/light/veml7700.c
@@ -206,7 +206,7 @@ static ssize_t device_read(struct file *file,   /* see include/linux/fs.h   */
 
 	printk(KERN_DEBUG "VEML7700 ######################################### LUX: %d * %f * %f = \n",ret,factor1,factor2);
 	//float buf = (float) ret;
-	ret = ret * 0.0288f;
+	ret = (int)(ret * 0.0288f);
 	printk(KERN_DEBUG "VEML7700 ######################################### RET: %d\n",ret);
